// frontend/src/generated/types/commonDetail.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface CommonDetail {
  sysId?: string;
  linkSysId: string;
  masterCommonCd?: string;
  commonCd?: string;
  commonNm: string;
  ordNo: number;
  useYn?: string;
}