// frontend/src/generated/types/commonDetailRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { CommonDetail } from './commonDetail';

export interface CommonDetailRequest {
  linkSysId?: string;
  newItems?: CommonDetail[];
  updateItems?: CommonDetail[];
  deleteItems?: CommonDetail[];
}